requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
loguru==0.7.0
pydantic==1.10.7
python-baseconv==1.2.2
python-ulid==1.1.0
sniffio==1.3.0
starlette==0.26.1
svix-ksuid==0.6.0
typing_extensions==4.5.0
uvicorn==0.22.0
